worked on fixing slider aswell as new wind breaker beat map
2rcd2rfaf

+++ JM -Shapes- Prototype/Assets/Scripts/AudioManager.cs
@@ -6,7 +6,7 @@ public class AudioManager : MonoBehaviour
     public GameObject beatManagers;
     private AudioSource audioSource;
     public AudioClip soundToPlay;
-    public int CountdownTime;
+    public float CountdownTime;
     void Awake()
     {
         audioSource = GetComponent<AudioSource>();
